Add scripts to register/unregister a build to RHSM
RHEL 7.6 builds failed on the cleanup script because they couldn't
install the package yum-utils due to there being no enabled
repos, because the system was not subscribed.

This commit adds a registration script which runs before the cleanup
script that subscribes the system to RHSM, and an unregistration
script that unregisters the system after the cleanup script. Since
you need to login to be able to subscribe the system, the commit
also adds the variables RHSM_USERNAME and RHSM_PASSWORD which will
be used by the script to authenticate against Red Hat during build
time.

Since the system is unregistered during the build, vagrant boxes will
need to be registered before any yum packages can be installed.